Main to Gwynn branch

+++ src/pages/register.js
@@ -1,13 +1,13 @@
-
-import { useState } from 'react';
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom'; // Assuming you are using react-router
 import { useUser } from '../hooks/useUser';
 import { Toaster } from 'sonner';
-import { useNavigate } from 'react-router-dom';
-import Navigation from '../components/Navbar'; 
+import Navigation from '../components/Navbar';
+import { toast } from 'sonner'
 
 const Signup = () => {
     const navigate = useNavigate();
-    const { errors, errorMessage, signUpUser } = useUser();
+    const { errors, errorMessage, signUpUser,  } = useUser();
 
     const [formData, setFormData] = useState({
         email: '',
@@ -17,38 +17,59 @@ const Signup = () => {
 
     const handleInputChange = (e) => {
         const { id, value } = e.target;
-        setFormData((prevData) => ({
-            ...prevData,
-            [id]: value,
-        }));
+        setFormData((prevData) => ({ ...prevData, [id]: value }));
     };
 
     const handleSignUpAndSave = async (e) => {
         e.preventDefault();
-    
-        try {
-          await signUpUser(formData);
-          // Store the email in local storage
-          localStorage.setItem('userEmail', formData.email);
-          // Redirect to the dashboard after successful signup
-          navigate('/dashboard');
-        } catch (error) {
-          console.error('Signup failed:', error);
+
+        // Additional validation checks can be added here
+
+        const { email, password, confirmPassword } = formData;
+        
+
+        // Check if any of the required fields are empty
+        if (!email || !password || !confirmPassword) {
+        console.error('Please fill in all required fields');
+        toast.error('Signup failed. Fill the Form.');
+        return;
+    }
+
+        // Check if password and confirmPassword match
+        if (password !== confirmPassword) {
+            // Set an error message
+            // You might want to display this error message in your UI
+            console.error("Passwords don't match");
+            toast.error('Signup failed. Please try again.');
+            return;
         }
-      };
+
+        try {
+            // Call the signup function from useUser hook
+            await signUpUser({ email, password });
+            
+            localStorage.setItem('userEmail', formData.email);
+            navigate('/dashboard');
+            toast.success('Account successfully registered!');
+            return true;
+
+        } catch (error) {
+            // Handle signup error
+            console.error('Signup failed:', error.message);
+        }
+    };
     
     return (
         <main>
-           <Navigation />
+            <Navigation />
             <Toaster richColors position="bottom-right" />
             <div className="div-container1">
-                
-                    
-                    {
-                        errorMessage &&
-                        <div className="text-red-500 mb-4 border-2 border-red-300 p-4 bg-red-200 rounded-md">{errorMessage}</div>
-                    }
-                    <form onSubmit={handleSignUpAndSave }>
+                {errorMessage && (
+                    <div className="text-red-500 mb-4 border-2 border-red-300 p-4 bg-red-200 rounded-md">
+                        {errorMessage}
+                    </div>
+                )}
+                <form onSubmit={handleSignUpAndSave}>
                     <h2 className="head-title">Sign Up</h2>
                         <div className="outerBox">
                             <label htmlFor="email" className="block text-gray-700 text-sm font-bold mb-2">
